Resolver Cache should load on-demand
Let's change it so we will only load the resolver cache when it is actually used. If a customer doesn't use PR functionality at all, it would never light up, so we save the bandwidth, CPU and memory.

Compared to 0.1.8, this change yields ~80% memory savings on a machine with lots of graph objects.